+++ phypidaq/MAX31855Config.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function, division, unicode_literals
-from __future__ import absolute_import
-
 import sys
 
 # import relevant pieces for MAX31855
-import Adafruit_GPIO.SPI as SPI
-import Adafruit_MAX31855.MAX31855 as MAX31855
+import board
+import digitalio
+import adafruit_max31855
 
 
 class MAX31855Config(object):
@@ -16,16 +14,10 @@ class MAX31855Config(object):
         if confdict is None:
             confdict = {}
 
-        # -- number of Channels
+        # Set number of Channels
         self.NChannels = 1
 
-        # -- configuration of chip select pin (CE0 or CE1)
-        if "CE" in confdict:
-            self.ce = confdict["CE"]
-        else:
-            self.ce = 0
-
-        # -- unit configuration MAX31855
+        # Set up unit configuration for MAX31855
         if "Unit" in confdict:
             if confdict["Unit"] == "DEGREES_F":
                 self.unit = confdict["Unit"]
@@ -37,30 +29,42 @@ class MAX31855Config(object):
                 self.unit = "DEGREES_C"
                 self.ChanLims = [[-10., 250.]]
         else:
+            # Default to Celsius
             self.unit = "DEGREES_C"
             self.ChanLims = [[-10., 250.]]
-
-    def init(self):
-        # Hardware configuration:
-        try:
-            # Create an MAX31855 instance.
-            self.MAX31855 = MAX31855.MAX31855(spi=SPI.SpiDev(0, self.ce))
-        except Exception as e:
-            print("MAX31855Config: Error initialising device - exit")
-            print(str(e))
-            sys.exit(1)
 
         # provide configuration parameters
         self.ChanNams = ['MAX31855']
 
+        # Set up hardware properties
+        self.spi = None
+        self.cs = None
+        self.sensor = None
+
+    def init(self):
+
+        # Init SPI and digital board IO
+        self.spi = board.SPI()
+        self.cs = digitalio.DigitalInOut(board.D5)
+
+        # Hardware configuration:
+        self.sensor = adafruit_max31855.MAX31855(self.spi, self.cs)
+
+        # Check if the configuration was successful
+        if self.sensor is None:
+            raise AttributeError("Couldn't initialize sensor!")
+
     def acquireData(self, buf):
-        if self.unit == "KELVIN":  # temperature in Kelvin
-            buf[0] = self.MAX31855.readTempC() + 273.15
-        elif self.unit == "DEGREES_F":  # temperature in degrees Fahrenheit
-            buf[0] = (self.MAX31855.readTempC() * 1.8) + 32
+        if self.unit == "KELVIN":
+            # Temperature in Kelvin
+            buf[0] = self.sensor.temperature + 273.15
+        elif self.unit == "DEGREES_F":
+            # Temperature in degrees Fahrenheit
+            buf[0] = (self.sensor.temperature * 1.8) + 32
         else:
-            buf[0] = self.MAX31855.readTempC()  # temperature in degrees Celsius
+            # Temperature in degrees Celsius
+            buf[0] = self.sensor.temperature
 
     def closeDevice(self):
-        # nothing to do here
+        # Nothing to do here
         pass